netDxf 2.0.1 Release
+ Implemented the code "\X" for dimension UserText. See sample DimensionUserTextWithTwoLines().
+ Rearrange the arguments in the OrdinateDimension constructor to maintain the same order in all of them.
+ (fixed) Issue Id #13525 HatchPattern now requires name.
+ (fixed) Issue Id #13524 Bug in DimensionStyleOverrideType.DIMDEC processing.

+++ netDxf/Tables/DimensionStyleOverride.cs
@@ -165,9 +165,9 @@ namespace netDxf.Tables
                         throw new ArgumentOutOfRangeException(nameof(value), value, string.Format("The {0} dimension style override must be greater than -1.", type));
                     break;
                 case DimensionStyleOverrideType.LengthPrecision:
-                    if (!(value is double))
-                        throw new ArgumentException(string.Format("The DimensionStyleOverrideType.{0} dimension style override must be a valid {1}", type, typeof (double)), nameof(value));
-                    if ((double) value < 0)
+                    if (!(value is short))
+                        throw new ArgumentException(string.Format("The DimensionStyleOverrideType.{0} dimension style override must be a valid {1}", type, typeof (short)), nameof(value));
+                    if ((short) value < 0)
                         throw new ArgumentOutOfRangeException(nameof(value), value, string.Format("The {0} dimension style override must be equals or greater than zero.", type));
                     break;
                 case DimensionStyleOverrideType.DimPrefix:
